+++ bittytax/price/valueasset.py
@@ -9,7 +9,6 @@ from tqdm import tqdm
 
 from ..version import __version__
 from ..config import config
-from ..tax import which_tax_year
 from .pricedata import PriceData
 
 class ValueAsset(object):
@@ -42,11 +41,11 @@ class ValueAsset(object):
                     config.CCY,
                     Style.NORMAL))
             return value, False
-        else:
-            tqdm.write("%sWARNING%s Price for %s on %s is not available, using price of %s" % (
-                Back.YELLOW+Fore.BLACK, Back.RESET+Fore.YELLOW,
-                asset, timestamp.strftime('%Y-%m-%d'), config.sym() + '{:0,.2f}'.format(0)))
-            return Decimal(0), False
+
+        tqdm.write("%sWARNING%s Price for %s on %s is not available, using price of %s" % (
+                   Back.YELLOW+Fore.BLACK, Back.RESET+Fore.YELLOW,
+                   asset, timestamp.strftime('%Y-%m-%d'), config.sym() + '{:0,.2f}'.format(0)))
+        return Decimal(0), False
 
     def get_current_value(self, asset, quantity):
         asset_price_ccy, name, data_source = self.get_latest_price(asset)
@@ -103,7 +102,10 @@ class ValueAsset(object):
 
     def price_report_cache(self, asset, timestamp, name, data_source, url,
                            price_ccy, price_btc=None):
-        tax_year = which_tax_year(timestamp)
+        if timestamp > config.get_tax_year_end(timestamp.year):
+            tax_year = timestamp.year + 1
+        else:
+            tax_year = timestamp.year
 
         if tax_year not in self.price_report:
             self.price_report[tax_year] = {}