+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Important:- A new Note field has been added to the end of the transaction record
 - Accounting tool: new config "transfer_fee_disposal" added (transfers_include=False). ([#56](https://github.com/BittyTax/BittyTax/issues/56))
 - Accounting tool: Excel files with worksheet names prefixed with '--' are ignored by the import.
 - Accounting tool: tax rates and allowance for 2021/22.
+- Accounting tool: tax rules for UK companies.
+- Accounting tool: tax rules option (--taxrules) added.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
 - Binance parser: use filename to determine if deposits or withdrawals.